bullet fixes

+++ Assets/Scripts/BubbleSpirit/BubbleBulletPatternLinear.cs
@@ -13,7 +13,7 @@ public class BubbleBulletPatternLinear : BubbleBulletPattern
 
         bullet.velocity = new Vector3(Mathf.Cos(angle),
                                       Mathf.Sin(angle),
-                                       0) * velocityParameters[0];
+                                      0) * (float)velocityParameters[0];
         bullet.angularVelocity = angularVelocity;
         //could use vector for acceleration too ...
         bullet.acceleration = acceleration;
